First steps in encapsulating model state.
Creation of qg_model_type to encapsulate psi, psit, qprime, and for.

Modify psitoq to operate on passed arguments instead of global data.

Modify qtopsi to operate on passed arguments instead of global data.

Modify ddt to operate on passed arguments instead of global data.

Modify artiforc to be a function that returns the forcing without
modifying any global data