Add all Tapenade adjoint routines to module_QG_adj.f90 and module_QG_ggsp.f90 except for ggtosp and sptogg.  This code will not compile because those are missing.  However, it does not make sense to call the Tapenade generated versions of those since they are unrelated to the derivative.  The fmtofs and fstofm also do not affect the derivative, but were included for expediency since they were relatively simple.
The push/pop routines are in the repository and are being called
for now, but are not linked in.

This commit is a place saver to ensure work is not lost.  Now
we must determine how all of the conversion routines (qtopsi,
psitoq, fmtofs, fstofm, ggtosp, sptogg) should be handled.  Also,
removal of as many (ideally, all of them) push/pop calls as possible
needs to be done as well.