Add Tapenade versions of ggtosp and sptogg but using the inverse FFT operations in place of the ones Tapenade generated.
Restore original Tapenade calls to the ggtosp and sptogg routines
for now.

Update the adjoint trajectory before the adjoint state.

This commit brings the adjoint to the following state.  The QGTLAD
test (with increment=1) shows perfect agreement between the state
of the ADJ at step t and the state of the TL at state t+1.

Efforts to line up the ADJ step to the TL step were unsuccessful so
far.  But, clearly most everything is working correctly.  There is
just some sort of phase error that is causing the TL and ADJ to be
one step off from each other.